pyproject.toml
[project]
name = "vton"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Julen Urchoeguia", email = "julen.urchoeguia@gmail.com" },
    { name = "Laure", email = "laureemilie.martin@gmail.com" }
]
dependencies = [
    "torch>=2.1.1",
    "pillow>=10.1.0",
    "numpy>=1.24.4",
    "ipykernel>=6.26.0",
    "torchvision>=0.16.1",
    "matplotlib>=3.7.3",
    "datasets>=2.15.0",
    "bitsandbytes>=0.41.3.post2",
    "refiners[training]>=0.2.0",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
